Add support for roof:orientation on skillion roofs
